// src/app/shared/models/Config.ts
export interface Config {
    id: string;
    name: string;
    memoryid: string;
    memoryname: string;
    motherboardid: string;
    motherboardname: string,
    powersupplyid: string;
    powersupplyname: string;
    processorid: string;
    processorname: string;
    videocardid: string;
    videocardname: string;
}